removed about-me and contact-me subpages
