--- pages/projects/index.module.scss
@@ -1,4 +1,0 @@
-// .projects {
-//   display: flex;
-//   justify-content: center;
-// }
